+++ src/prolog/mod.rs
@@ -1,11 +1,8 @@
-use self::subset::subset::Subset;
-
 pub mod content_particle;
 pub mod declaration_content;
 pub mod doctype;
 pub mod external_id;
 pub mod id;
-// TODO: pub mod internal_subset;
 pub mod subset;
 pub mod textdecl;
 pub mod xmldecl;
